OSAC-4750: split sanity/regression tests for CaaS
