+++ tests/e2e/caas/regression/test_cluster_version.py
@@ -1,0 +1,106 @@
+from __future__ import annotations
+
+import contextlib
+import subprocess
+from pathlib import Path
+
+from tests.e2e.catalog.conftest import unique_name
+from tests.e2e.core.grpc_client import GRPCClient
+from tests.e2e.core.helpers import (
+    wait_for_cluster_deleting,
+    wait_for_cluster_deletion,
+    wait_for_cluster_grpc_deleting_or_archived,
+    wait_for_cluster_grpc_removal,
+    wait_for_cluster_order_cr,
+)
+from tests.e2e.core.k8s_client import K8sClient
+from tests.e2e.core.osac_cli import OsacCLI
+from tests.e2e.core.runner import poll_until
+
+# Fixed so repeated runs reuse the same ClusterVersion on the shared cluster.
+TEST_RELEASE_IMAGE = "quay.io/openshift-release-dev/ocp-release:4.20.0-multi"
+
+
+def test_cluster_create_with_version(
+    cli: OsacCLI,
+    grpc: GRPCClient,
+    private_grpc: GRPCClient,
+    k8s_hub_client: K8sClient,
+    cluster_template: str,
+    pull_secret_path: str,
+    ssh_public_key_path: str,
+) -> None:
+    """Verify explicit version resolution and reference protection."""
+    version = private_grpc.ensure_cluster_version(version="4.20.0-e2e", image=TEST_RELEASE_IMAGE)
+
+    name = unique_name("e2e-cluster-version")
+    uuid = cli.create_cluster(
+        name=name,
+        template=cluster_template,
+        version=version["name"],
+        template_parameter_files={"pull_secret": pull_secret_path},
+        template_parameters={"ssh_public_key": Path(ssh_public_key_path).read_text().strip()},
+    )
+
+    try:
+        co_name = wait_for_cluster_order_cr(k8s=k8s_hub_client, uuid=uuid)
+
+        cluster = grpc.get_cluster(cluster_id=uuid)
+        assert cluster["object"]["spec"]["version"]["name"] == version["name"]
+
+        # While the cluster references this version, deletion must be rejected.
+        output, rc = private_grpc.call_unchecked(
+            service="osac.private.v1.ClusterVersions/Delete", data={"id": version["id"]}
+        )
+        assert rc != 0, f"Expected delete to be rejected for referenced version, got: {output}"
+        assert "FailedPrecondition" in output or "referenced" in output.lower(), (
+            f"Expected FailedPrecondition or 'referenced' in rejection, got: {output}"
+        )
+
+        release_image = poll_until(
+            fn=lambda: k8s_hub_client.get_cluster_order_spec(name=co_name).get("releaseImage", ""),
+            until=lambda v: v != "",
+            retries=30,
+            delay=5,
+            description=f"{co_name} ClusterOrder releaseImage resolution",
+        )
+        assert release_image == TEST_RELEASE_IMAGE
+
+        cli.delete_cluster(uuid=uuid)
+
+        wait_for_cluster_deleting(k8s=k8s_hub_client, name=co_name)
+        wait_for_cluster_grpc_deleting_or_archived(grpc=grpc, uuid=uuid)
+        wait_for_cluster_deletion(k8s=k8s_hub_client, name=co_name)
+        wait_for_cluster_grpc_removal(grpc=grpc, uuid=uuid)
+    finally:
+        with contextlib.suppress(subprocess.CalledProcessError):
+            cli.delete_cluster(uuid=uuid)
+
+
+def test_cluster_create_rejected_for_invalid_version(
+    grpc: GRPCClient, private_grpc: GRPCClient, cluster_template: str
+) -> None:
+    """Verify creation is rejected for disabled, obsolete, and missing versions."""
+    disabled = private_grpc.ensure_cluster_version(version="4.20.0-e2e-disabled", image=TEST_RELEASE_IMAGE)
+    private_grpc.update_cluster_version(version_id=disabled["id"], enabled=False)
+
+    obsolete = private_grpc.ensure_cluster_version(version="4.20.0-e2e-obsolete", image=TEST_RELEASE_IMAGE)
+    private_grpc.update_cluster_version(version_id=obsolete["id"], state="CLUSTER_VERSION_STATE_OBSOLETE")
+
+    def _create_with_version(version_name: str) -> tuple[str, int]:
+        return grpc.call_unchecked(
+            service="osac.public.v1.Clusters/Create",
+            data={"object": {"spec": {"template": {"name": cluster_template}, "version": {"name": version_name}}}},
+        )
+
+    output, rc = _create_with_version(disabled["name"])
+    assert rc != 0, f"Expected create to reject disabled version, got: {output}"
+    assert "disabled" in output.lower(), f"Expected 'disabled' in rejection, got: {output}"
+
+    output, rc = _create_with_version(obsolete["name"])
+    assert rc != 0, f"Expected create to reject obsolete version, got: {output}"
+    assert "obsolete" in output.lower(), f"Expected 'obsolete' in rejection, got: {output}"
+
+    output, rc = _create_with_version("4-20-0-e2e-does-not-exist")
+    assert rc != 0, f"Expected create to reject non-existent version, got: {output}"
+    assert "not found" in output.lower(), f"Expected 'not found' in rejection, got: {output}"
