+++ tests/e2e/caas/sanity/test_cluster_delete_feedback_light.py
